change car type to int in OrderItem class

+++ src/main/java/com/islamhamada/petshop/entity/OrderItem.java
@@ -22,5 +22,5 @@ public class OrderItem {
 
     private long productId;
 
-    private long count;
+    private int count;
 }
